+++ tests/core/mod.rs
@@ -1,6 +1,7 @@
 #[cfg(feature = "abomonation-serialize")]
 mod abomonation;
 mod blas;
+mod cg;
 mod conversion;
 mod edition;
 mod empty;
@@ -9,6 +10,9 @@ mod matrix_slice;
 #[cfg(feature = "mint")]
 mod mint;
 mod serde;
+
+#[cfg(feature = "compare")]
+mod matrixcompare;
 
 #[cfg(feature = "arbitrary")]
 pub mod helper;